add pipenv e app pet
