+++ README.txt
@@ -1,17 +1,5 @@
-pip freeze > requirements.txt (atualizar dependencias da aplicação)
-
-pip install -r requirements.txt (instalar dependencias)
-
-Rodar o Django:
-
-1. fazer migraçoes necessarias
-python manage.py makemigrations
-python manage.py migrate
-
-2. rodar o server
-python manage.py runserver
-
-3. acessar os urls abaixo
+PRIMEIRO DE TUDO - rodar o virtual environment:
+pipenv shell
 
 API urls:
 
@@ -23,3 +11,17 @@ http://127.0.0.1:8000/api/users/3/
 
 LOGIN:
 http://127.0.0.1:8000/api-token-auth/
+
+Rodar o Django:
+
+1. fazer migraçoes necessarias
+python manage.py makemigrations
+python manage.py migrate
+
+2. rodar o server
+python manage.py runserver
+
+3. acessar os urls acima
+
+Criar novo app:
+python manage.py startapp <name of app>
