--- backend/accounts/readme.txt
@@ -1,10 +1,0 @@
-API urls:
-
-CRUD:
-http://127.0.0.1:8000/api/users/
-
-Para acessar usuario especifico: (exemplo)
-http://127.0.0.1:8000/api/users/3/
-
-LOGIN:
-http://127.0.0.1:8000/api-token-auth/
