+++ backend/core/admin.py
@@ -1,5 +1,1 @@
 from django.contrib import admin
-from .models import PET
-
-# Register your models here.
-admin.site.register(PET)
