+++ backend/pet/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+
+# Register your models here.
+from .models import PET
+
+admin.site.register(PET)
